# target/generated-sources/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  description: Data Platform User Context Service
  title: user-context-service
  version: v1
servers:
- url: /
...
paths:
  /contexts:
    get:
      description: User must have LIST_CONTEXTS permission
      operationId: getAllContexts
      parameters:
      - description: What page to grab
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - description: Number of elements on page
        explode: true
        in: query
        name: pageSize
        required: false
        schema:
          default: 25
          type: integer
        style: form
      - description: Sort order direction
        explode: true
        in: query
        name: direction
        required: false
        schema:
          default: ASC
          enum:
          - ASC
          - DESC
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextPage'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: gets all contexts
      tags:
      - Contexts
      x-accepts: application/json
    post:
      description: User must have CREATE_CONTEXTS permission
      operationId: createContext
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Context'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: creates new context via upsert
      tags:
      - Contexts
      x-accepts: application/json
      x-content-type: application/json
  /contexts/{contextId}:
    delete:
      description: User must have DELETE_CONTEXTS permission
      operationId: deleteContextByContextId
      parameters:
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: deleted successfully
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: deletes context by contextId
      tags:
      - Contexts
      x-accepts: application/json
    get:
      description: User must have LIST_CONTEXTS permission
      operationId: getContextByContextId
      parameters:
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: gets context by contextId
      tags:
      - Contexts
      x-accepts: application/json
    put:
      description: User must have UPDATE_CONTEXTS permission
      operationId: updateContextByContextId
      parameters:
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Context'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: updates context by contextId
      tags:
      - Contexts
      x-content-type: application/json
      x-accepts: application/json
  /contexts/{contextId}/roles:
    get:
      description: User must have LIST_CONTEXTS permission
      operationId: getRolesByContextId
      parameters:
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePage'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: get roles by contextId
      tags:
      - Contexts
      x-accepts: application/json
    post:
      description: User must have UPDATE_CONTEXTS permission
      operationId: addAdditionalRolesByContextId
      parameters:
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: get roles by contextId
      tags:
      - Contexts
      x-content-type: application/json
      x-accepts: application/json
  /contexts/{contextId}/roles/{roleId}:
    get:
      description: User must have LIST_CONTEXTS permission
      operationId: getRoleByContextIdAndRoleId
      parameters:
      - description: contextId
        explode: false
        in: path
        name: contextId
        required: true
        schema:
          type: string
        style: simple
      - description: roleId
        explode: false
        in: path
        name: roleId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control header provides important info on how long
                a response may be considered fresh
              example: "Cache-Control: max-age=3600, public, must-revalidate"
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      summary: get context by contextId and roleId
      tags:
      - Contexts
      x-accepts: application/json
...
components:
  schemas:
...
    ContextPage:
      example:
        elapsed: elapsed
        total: 1
        pageSize: 6
        correlationId: correlationId
        page: 0
        content:
        - contextName: contextName
          modifiedTime: modifiedTime
          createdBy: createdBy
          permissions:
          - permissions
          - permissions
          roles:
          - roleId: roleId
            permissions:
            - permissions
            - permissions
            roleName: roleName
          - roleId: roleId
            permissions:
            - permissions
            - permissions
            roleName: roleName
          description: description
          createdTime: createdTime
          contextId: contextId
          modifiedBy: modifiedBy
          id: id
          enabled: true
          transactionId: transactionId
        - contextName: contextName
          modifiedTime: modifiedTime
          createdBy: createdBy
          permissions:
          - permissions
          - permissions
          roles:
          - roleId: roleId
            permissions:
            - permissions
            - permissions
            roleName: roleName
          - roleId: roleId
            permissions:
            - permissions
            - permissions
            roleName: roleName
          description: description
          createdTime: createdTime
          contextId: contextId
          modifiedBy: modifiedBy
          id: id
          enabled: true
          transactionId: transactionId
      properties:
        content:
          items:
            $ref: '#/components/schemas/Context'
          type: array
        elapsed:
          type: string
        correlationId:
          type: string
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      type: object
    Context:
      example:
        contextName: contextName
        modifiedTime: modifiedTime
        createdBy: createdBy
        permissions:
        - permissions
        - permissions
        roles:
        - roleId: roleId
          permissions:
          - permissions
          - permissions
          roleName: roleName
        - roleId: roleId
          permissions:
          - permissions
          - permissions
          roleName: roleName
        description: description
        createdTime: createdTime
        contextId: contextId
        modifiedBy: modifiedBy
        id: id
        enabled: true
        transactionId: transactionId
      properties:
        id:
          type: string
        contextId:
          type: string
        contextName:
          type: string
        description:
          type: string
        permissions:
          items:
            type: string
          minItems: 1
          type: array
          uniqueItems: true
        roles:
          items:
            $ref: '#/components/schemas/Role'
          minItems: 1
          type: array
        enabled:
          type: boolean
        transactionId:
          type: string
        createdBy:
          type: string
        modifiedBy:
          type: string
        createdTime:
          type: string
        modifiedTime:
          type: string
      required:
      - contextId
      - contextName
      - description
      - permissions
      - roles
      type: object
    RolePage:
      example:
        elapsed: elapsed
        total: 1
        pageSize: 6
        correlationId: correlationId
        page: 0
        content:
        - roleId: roleId
          permissions:
          - permissions
          - permissions
          roleName: roleName
        - roleId: roleId
          permissions:
          - permissions
          - permissions
          roleName: roleName
      properties:
        content:
          items:
            $ref: '#/components/schemas/Role'
          type: array
        elapsed:
          type: string
        correlationId:
          type: string
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      type: object
    Role:
      example:
        roleId: roleId
        permissions:
        - permissions
        - permissions
        roleName: roleName
      properties:
        roleId:
          type: string
        roleName:
          type: string
        permissions:
          items:
            type: string
          minItems: 1
          type: array
          uniqueItems: true
      required:
      - permissions
      - roleId
      - roleName
      type: object
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http